requirements.txt
langchain==1.2.10
langgraph==1.0.8
requests==2.32.5
langchain_groq==1.1.2
python-dotenv==1.0.0
duckduckgo-search==8.1.1
langchain-community==0.4.1
ddgs==9.10.0
deep-translator==1.11.4
ipython==9.10.0